requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
bcrypt==5.0.0
Django==6.0.2
django-extensions==4.1
django-ninja==1.5.3
ecdsa==0.19.1
passlib==1.7.4
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
